refactor where availability component is located
