+++ components/Home/ManageAvailabilityWizard/AvailabilityModal.tsx
@@ -7,11 +7,11 @@ import ClipLoader from "react-spinners/ClipLoader";
 import { DayRow } from "./DayRow";
 import { DateTime } from "luxon";
 import useSWR from "swr";
-import { APIContext } from "../../context/APIContext";
-import { Loader } from "../util/Loader";
-import { CustomError } from "../util/CustomError";
-import { Availability } from "../../models/availability";
-import { ValidDays } from "../../models/availability";
+import { APIContext } from "../../../context/APIContext";
+import { Loader } from "../../util/Loader";
+import { CustomError } from "../../util/CustomError";
+import { Availability } from "../../../models/availability";
+import { ValidDays } from "../../../models/availability";
 import axios from "axios";
 
 type AvailabilityModalProps = {
