+++ components/Home/ManageAvailabilityWizard/TimeSlot.module.css
@@ -3,12 +3,12 @@
   border-radius: 8px;
   background: var(--gray-4);
   color: var(--black);
-  composes: labelLarge from "../../styles/globals.css";
+  composes: labelLarge from "../../../styles/globals.css";
 }
 
 .timeInput {
   composes: inputField;
-  composes: labelLarge from "../../styles/globals.css";
+  composes: labelLarge from "../../../styles/globals.css";
   width: 85px;
   height: 40px;
   margin: 10px 0px;
